requirements.txt
mysql-connector-python
flask
beautifulsoup4
requests
scikit-learn